fix: relocate top level files
