+++ src/public/index.js
@@ -1,9 +1,9 @@
 import * as React from "react";
 import * as ReactDOM from "react-dom";
 import { ChakraProvider, ColorModeScript } from "@chakra-ui/react";
-import App from "./components/App";
+import App from "../components/App";
 import "./index.css";
-import theme from "./theme";
+import theme from "../ui/theme";
 
 ReactDOM.render(
   <>
